alarmed events and simple graphs
Previously there was Events and Alarms. However there isn't much data
stored in Alarms. A new alarmed flag was added to Events and that
got rid of the Alarm table.

Cleared up the Severity versus EventType. Now Severity is a property
of EventState and is just a convenience for its color. EventType
does not have anything to do with colors.

Realised I was typing the same thing over and over for most graphs
so invented simple graphs. These have a title, some fields and the
whole line stuff is done with 4 fields. Automatic avg/max/last too.