Using standard grids and added info socket
rnmsd would loop tightly as it had no sockets to listen on, so was going
to add info socket so here it is.

Most but not all of the grids now use the standard TableFiller and
TableBase, admin grids all need re-work for their special grids with
edit and delete items.