OIDs use tuples and fixed admin
When using the with_oid flag for SNMP queries, the result is
now a oid,value tuple. This means that for things that just need
to scan the list can and others can use dict()

Admin screens have a working delete button and the grids behave
themselves. Will need to fix the non-admin grids the same way now too.